--- mcp_tools/network_recon/enhanced_network_scanning.py
@@ -1,147 +1,0 @@
-# mcp_tools/enhanced_network_scanning.py
-
-from typing import Dict, Any
-
-def register_enhanced_network_scanning_tools(mcp, hexstrike_client, logger):
-    @mcp.tool()
-    def rustscan_fast_scan(target: str, ports: str = "", ulimit: int = 5000,
-                          batch_size: int = 4500, timeout: int = 1500,
-                          scripts: bool = False, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Rustscan for ultra-fast port scanning with enhanced logging.
-
-        Args:
-            target: The target IP address or hostname
-            ports: Specific ports to scan (e.g., "22,80,443")
-            ulimit: File descriptor limit
-            batch_size: Batch size for scanning
-            timeout: Timeout in milliseconds
-            scripts: Run Nmap scripts on discovered ports
-            additional_args: Additional Rustscan arguments
-
-        Returns:
-            Ultra-fast port scanning results
-        """
-        data = {
-            "target": target,
-            "ports": ports,
-            "ulimit": ulimit,
-            "batch_size": batch_size,
-            "timeout": timeout,
-            "scripts": scripts,
-            "additional_args": additional_args
-        }
-        logger.info(f"⚡ Starting Rustscan: {target}")
-        result = hexstrike_client.safe_post("api/tools/rustscan", data)
-        if result.get("success"):
-            logger.info(f"✅ Rustscan completed for {target}")
-        else:
-            logger.error(f"❌ Rustscan failed for {target}")
-        return result
-
-    @mcp.tool()
-    def masscan_high_speed(target: str, ports: str = "1-65535", rate: int = 1000,
-                          interface: str = "", router_mac: str = "", source_ip: str = "",
-                          banners: bool = False, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Masscan for high-speed Internet-scale port scanning with intelligent rate limiting.
-
-        Args:
-            target: The target IP address or CIDR range
-            ports: Port range to scan
-            rate: Packets per second rate
-            interface: Network interface to use
-            router_mac: Router MAC address
-            source_ip: Source IP address
-            banners: Enable banner grabbing
-            additional_args: Additional Masscan arguments
-
-        Returns:
-            High-speed port scanning results with intelligent rate limiting
-        """
-        data = {
-            "target": target,
-            "ports": ports,
-            "rate": rate,
-            "interface": interface,
-            "router_mac": router_mac,
-            "source_ip": source_ip,
-            "banners": banners,
-            "additional_args": additional_args
-        }
-        logger.info(f"🚀 Starting Masscan: {target} at rate {rate}")
-        result = hexstrike_client.safe_post("api/tools/masscan", data)
-        if result.get("success"):
-            logger.info(f"✅ Masscan completed for {target}")
-        else:
-            logger.error(f"❌ Masscan failed for {target}")
-        return result
-
-    @mcp.tool()
-    def rpcclient_enumeration(target: str, username: str = "", password: str = "",
-                             domain: str = "", commands: str = "enumdomusers;enumdomgroups;querydominfo",
-                             additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute rpcclient for RPC enumeration with enhanced logging.
-
-        Args:
-            target: The target IP address
-            username: Username for authentication
-            password: Password for authentication
-            domain: Domain for authentication
-            commands: Semicolon-separated RPC commands
-            additional_args: Additional rpcclient arguments
-
-        Returns:
-            RPC enumeration results
-        """
-        data = {
-            "target": target,
-            "username": username,
-            "password": password,
-            "domain": domain,
-            "commands": commands,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting rpcclient: {target}")
-        result = hexstrike_client.safe_post("api/tools/rpcclient", data)
-        if result.get("success"):
-            logger.info(f"✅ rpcclient completed for {target}")
-        else:
-            logger.error(f"❌ rpcclient failed for {target}")
-        return result
-
-    @mcp.tool()
-    def nbtscan_netbios(target: str, verbose: bool = False, timeout: int = 2,
-                       additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute nbtscan for NetBIOS name scanning with enhanced logging.
-
-        Args:
-            target: The target IP address or range
-            verbose: Enable verbose output
-            timeout: Timeout in seconds
-            additional_args: Additional nbtscan arguments
-
-        Returns:
-            NetBIOS name scanning results
-        """
-        data = {
-            "target": target,
-            "verbose": verbose,
-            "timeout": timeout,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting nbtscan: {target}")
-        result = hexstrike_client.safe_post("api/tools/nbtscan", data)
-        if result.get("success"):
-            logger.info(f"✅ nbtscan completed for {target}")
-        else:
-            logger.error(f"❌ nbtscan failed for {target}")
-        return result
-
-
-
-
-
-
